+++ inc/tp_serial_protocol/SimpleSerialInterfaceManager.h
@@ -75,7 +75,7 @@ public:
 
   //################################################################################################
   //! Number of bytes in the queue
-  int queueSize(const std::string& path)const;
+  size_t queueSize(const std::string& path)const;
 
   //################################################################################################
   //! This is called just before a serial port is opened
